bruno j

+++ MercadoPy/mercado.py
@@ -54,7 +54,7 @@ def cadastrar_produto() -> None:
     print('===================')
 
     nome: str = input('Informe o nome do produto: ')
-    preco: float = int(input('Informe o preço do produto: '))
+    preco: float = float(input('Informe o preço do produto: '))
 
     produto: Produto = Produto(nome, preco)
